Added Audio Desync Hack core option
Add core option to address the problem of audio desynchronizing when
NoWaitMode is enabled. It works by simply discarding any audio samples
generated past the requested amount per frame slice, which from my tests
isn't many and my ears can't notice a difference. Forces NoWaitMode to
enabled regardless of its setting, thus leaving speed throttling up to
the frontend by way of some combination of its various sync options
(video, audio, and/or exact content framerate/time sync) - which is "the
right way" for a libretro core.

AFAICT NoWaitMode set to disabled avoids this problem by periodically
oversleeping and thereby starving the audio buffer. This hack is no less
ugly of a fix, sure it's not the right way but neither is throttling in
the core in the way NoWaitMode disabled does, IMO. If/when the root
problem is addressed this option (and NoWaitMode) will no longer be
needed.